Fixed even-positioned values greater than 9 concatenating instead of adding up

+++ src/Luhn.php
@@ -31,7 +31,10 @@ function luhn($value = '')
 		// e..g, 3 => 6
 		$total += !$isEven
 			? $val
-			: ($buffer = (int) $val * 2) > 9 ? (string) $buffer[0] + (string) $buffer[1] :$buffer);
+			: ($buffer = ((int) $val) * 2) > 9
+				? (int) (((string) $buffer)[0]) + (int) (((string) $buffer)[1])
+				: $buffer
+			);
 
 		// Indicate that the next iteration will be an odd position
 		$isEven = !$isEven;
